+++ mne/preprocessing/ica.py
@@ -50,6 +50,7 @@ from ..filter import band_pass_filter
 from .bads import find_outliers
 from .ctps_ import ctps
 from ..externals.six import string_types, text_type
+from ..io.pick import channel_type
 
 
 __all__ = ['ICA', 'ica_find_ecg_events', 'ica_find_eog_events',
@@ -86,6 +87,27 @@ def get_score_funcs():
                             for n, f in xy_arg_stats_funcs
                             if _get_args(f) == ['x', 'y']))
     return score_funcs
+
+
+def _check_for_unsupported_ica_channels(picks, info):
+    """Check for channels in picks that are not considered
+    valid channels. Accepted channels are the data channels
+    ('seeg','ecog','eeg','mag', and 'grad') and 'eog'.
+    This prevents the program from crashing without
+    feedback when a bad channel is provided to ICA whitening.
+    """
+    if picks is None:
+        return
+    elif len(picks) == 0:
+        raise ValueError('No channels provided to ICA')
+    types = _DATA_CH_TYPES_SPLIT + ['eog']
+    chs = list(set([channel_type(info, j) for j in picks]))
+    check = all([ch in types for ch in chs])
+    if not check:
+        raise ValueError('Invalid channel type(s) passed for ICA.\n'
+                         'Only the following channels are supported {}\n'
+                         'Following types were passed {}\n'
+                         .format(types, chs))
 
 
 class ICA(ContainsMixin):
@@ -326,11 +348,13 @@ class ICA(ContainsMixin):
         self : instance of ICA
             Returns the modified instance.
         """
-        if isinstance(inst, _BaseRaw):
-            self._fit_raw(inst, picks, start, stop, decim, reject, flat,
-                          tstep, verbose)
-        elif isinstance(inst, _BaseEpochs):
-            self._fit_epochs(inst, picks, decim, verbose)
+        if isinstance(inst, _BaseRaw) or isinstance(inst, _BaseEpochs):
+            _check_for_unsupported_ica_channels(picks, inst.info)
+            if isinstance(inst, _BaseRaw):
+                self._fit_raw(inst, picks, start, stop, decim, reject, flat,
+                              tstep, verbose)
+            elif isinstance(inst, _BaseEpochs):
+                self._fit_epochs(inst, picks, decim, verbose)
         else:
             raise ValueError('Data input must be of Raw or Epochs type')
         return self
@@ -440,7 +464,7 @@ class ICA(ContainsMixin):
             # Scale (z-score) the data by channel type
             info = pick_info(info, picks)
             pre_whitener = np.empty([len(data), 1])
-            for ch_type in _DATA_CH_TYPES_SPLIT:
+            for ch_type in _DATA_CH_TYPES_SPLIT + ['eog']:
                 if _contains_ch_type(info, ch_type):
                     if ch_type == 'seeg':
                         this_picks = pick_types(info, meg=False, seeg=True)
@@ -448,8 +472,14 @@ class ICA(ContainsMixin):
                         this_picks = pick_types(info, meg=False, ecog=True)
                     elif ch_type == 'eeg':
                         this_picks = pick_types(info, meg=False, eeg=True)
-                    else:
+                    elif ch_type in ['mag', 'grad']:
                         this_picks = pick_types(info, meg=ch_type)
+                    elif ch_type == 'eog':
+                        this_picks = pick_types(info, meg=False, eog=True)
+                    else:
+                        raise RuntimeError('Should not be reached.'
+                                           'Unsupported channel {}'
+                                           .format(ch_type))
                     pre_whitener[this_picks] = np.std(data[this_picks])
             data /= pre_whitener
         elif not has_pre_whitener and self.noise_cov is not None: